+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Clean legacy code
 - Fix a bug that can lead to a QGIS crash because of legacy method
+- add a button to reset settings to factory default
 
 ## 1.0.3 - 2022-10-18